% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicsegmentation.R
\name{dynamicSegmentation_interface}
\alias{dynamicSegmentation_interface}
\alias{dynamicSegmentation_algo}
\title{Dynamic Segmentation Algorithm}
\usage{
dynamicSegmentation_algo(data)
}
\arguments{
\item{data}{The input numeric vector.}
}
\value{
An integer vector of changepoint indices.
}
\description{
This function performs dynamic segmentation on the input data.
}
\examples{
data <- c(1.0, 2.0, 3.0, 4.0, 5.0)
dynamicSegmentation_algo(data)
}
